refactor: split srcset extraction from image source module
